moves IPOPT callback to separate library
